kinesis test

+++ 06-best-practices/code/integraton-test/run.sh
@@ -5,6 +5,8 @@ cd "$(dirname "$0")"
 
 LOCAL_TAG=`date +"%Y-%m-%d-%H-%M"`
 export LOCAL_IMAGE_NAME="stream-model-duration:${LOCAL_TAG}"
+export PREDICTIONS_STREAM_NAME="ride_predictions"
+
 
 docker build -t ${LOCAL_IMAGE_NAME} ..
 
@@ -12,14 +14,31 @@ docker-compose up -d
 
 sleep 1
 
+aws --endpoint-url=http://localhost:4566 \
+    kinesis create-stream \
+    --stream-name ${PREDICTIONS_STREAM_NAME} \
+    --shard-count 1
+
 pipenv run python test_docker.py
 
 ERROR_CODE=$?
 
 if [ ${ERROR_CODE} != 0 ]; then
     docker-compose logs
+    docker-compose down
+    exit ${ERROR_CODE}
 fi
 
-docker-compose down
 
-exit ${ERROR_CODE}
+pipenv run python test_kinesis.py
+
+ERROR_CODE=$?
+
+if [ ${ERROR_CODE} != 0 ]; then
+    docker-compose logs
+    docker-compose down
+    exit ${ERROR_CODE}
+fi
+
+
+docker-compose down
